+++ src/plugins/medium-zoom-docusaurus-plugin/attach-medium-zoom.js
