+++ src/plugins/medium-zoom-docusaurus-plugin/index.js
